new params dictionary 'data' and modified demos'
